scrapers/requirements.txt
# Web scraping libraries
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
scrapy==2.11.0
playwright==1.40.0

# Database and API integration
supabase==2.1.0
pydantic==2.5.2
python-dotenv==1.0.0

# Utilities
apscheduler==3.10.4
loguru==0.7.2
tenacity==8.2.3

# Translation (for Japanese content)
deep-translator==1.11.4
